Move the core mujoco simulation to its own container
